Pin Kafka version
Regardless of what version of Kafka clients upstream dependencies bring in, we need the project to be at a known Kafka version so that the module patching works, i.e.

```kotlin
modularity.patchModule("kafka.streams", "kafka-streams-test-utils-$kafkaVersion.jar")
```

If we don't pin, this will break when creek-kafka is released with different version of Kafka jars.